PR-5 Refactor NavigationBar
